requirements.txt
django
djangorestframework
environs
requests
ollama